+++ smarts/zoo/manager.py
@@ -1,4 +1,6 @@
-# Copyright (C) 2020. Huawei Technologies Co., Ltd. All rights reserved.
+# MIT License
+#
+# Copyright (C) 2021. Huawei Technologies Co., Ltd. All rights reserved.
 #
 # Permission is hereby granted, free of charge, to any person obtaining a copy
 # of this software and associated documentation files (the "Software"), to deal
@@ -17,13 +19,11 @@
 # LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 # THE SOFTWARE.
+
 import argparse
 import logging
 import os
-import pathlib
 import signal
-import subprocess
-import sys
 from concurrent import futures
 
 import grpc
@@ -43,7 +43,7 @@ def serve(port):
     server.start()
     log.debug(f"Manager - ip({ip}), port({port}), pid({os.getpid()}): Started serving.")
 
-    def stop_server(unused_signum, unused_frame):
+    def stop_server(*args):
         manager_servicer_object.destroy()
         server.stop(0)
         log.debug(