+++ CHANGELOG.md
@@ -20,7 +20,8 @@ Copy and pasting the git commit messages is __NOT__ enough.
 - Added Base CI and dependencies requirement tests for the "darwin" platform (MacOS).
 - Extended Imitation Learning codebase to allow importing traffic histories from the Waymo motion dataset and replay in a SMARTS simulation. See PR #1060.
 - Added `ros` extension rule to `setup.py`.
-- Added a -y option to `utils/setup/install_deps.sh` to allow default accept of installation. See Issue #1081.
+- Added a `-y` option to `utils/setup/install_deps.sh` to accept installation by default. See issue #1081.
+- Added `ParallelEnv` class and a corresponding example to simulate multiple SMARTS environments in parallel, with synchronous or asynchronous episodes.
 ### Changed
 - Made changes to log sections of the scenario step in `smarts.py` to help evaluate smarts performance problems. See Issue #661.
 - Introducted `RoadMap` class to abstract away from `SumoRoadNetwork`